requirements.txt
streamlit
streamlit-drawable-canvas
Pillow
numpy
torch
torchvision
pandas
